LPS-198452 Move class CPDefinitionLinkTypeConfigurationWrapper to class CPDefinitionLinkTypeSettingsImpl
