--- modules/apps/commerce/commerce-product-service/src/main/java/com/liferay/commerce/product/internal/configuration/CPDefinitionLinkTypeConfigurationWrapper.java
@@ -1,43 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.commerce.product.internal.configuration;
-
-import com.liferay.commerce.product.configuration.CPDefinitionLinkTypeConfiguration;
-import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
-
-import java.util.Map;
-
-import org.osgi.service.component.annotations.Activate;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.ConfigurationPolicy;
-import org.osgi.service.component.annotations.Modified;
-
-/**
- * @author Alessio Antonio Rendina
- */
-@Component(
-	configurationPid = "com.liferay.commerce.product.configuration.CPDefinitionLinkTypeConfiguration",
-	configurationPolicy = ConfigurationPolicy.REQUIRE,
-	service = CPDefinitionLinkTypeConfigurationWrapper.class
-)
-public class CPDefinitionLinkTypeConfigurationWrapper {
-
-	public String getType() {
-		return _cpDefinitionLinkTypeConfiguration.type();
-	}
-
-	@Activate
-	@Modified
-	protected void activate(Map<String, Object> properties) {
-		_cpDefinitionLinkTypeConfiguration =
-			ConfigurableUtil.createConfigurable(
-				CPDefinitionLinkTypeConfiguration.class, properties);
-	}
-
-	private volatile CPDefinitionLinkTypeConfiguration
-		_cpDefinitionLinkTypeConfiguration;
-
-}
